refactor(engine-host): move auth-mode resolution below the seam
Move resolve_auth_mode / resolve_model_switch_auth_mode (+ the private
resolve_configured_auth_mode / _for_entry helpers) out of main.rs into
engine_host::config, beside the model/permission resolvers they sit with. They
resolve which auth mode a model uses from the config -- an engine input, not a
renderer concern -- and name only engine_core:: types (no core-cluster ref, no
cycle). main.rs imports the two public entry points; the auth_mode_tests stay
in the CLI bin (the module is a grab-bag that also tests CLI-only helpers) and
reach resolve_model_switch_auth_mode through the re-import via `use super::*`.

Fifth staged step (part a of c) of the engine-host extraction (#120).
